Migrate size type to int

+++ src/chapter15/e15_34/src/main/java/com/dawids/RandomWalkControl.java
@@ -7,9 +7,9 @@ public class RandomWalkControl {
         this(10);
     }
 
-    public RandomWalkControl(double size) {
-        head.setX((int) (size / 2));
-        head.setY((int) (size / 2));
+    public RandomWalkControl(int size) {
+        head.setX(size / 2);
+        head.setY(size / 2);
     }
 
     public void makeMove(Directions direction) {
